fix(tree-hash): file names
rename as permutation (4 1 3)
