--- docs/graph/config/tree-hash/tree-hash_1.yml
@@ -1,10 +1,0 @@
-id: oi-wiki:graph/tree-hash-1
-name: Tree Hash
-solution:
-  cpp: docs/graph/code/tree-hash/tree-hash_1.cpp
-target:
-- 'test'
-testcases:
-  answer: docs/graph/examples/tree-hash/tree-hash_1.ans
-  input: docs/graph/examples/tree-hash/tree-hash_1.in
-version: 1000
