+++ docs/graph/examples/tree-hash/tree-hash_1.ans
@@ -1,4 +1,1 @@
-1 
-1 
-3 
-1 
+4
