+++ docs/graph/examples/tree-hash/tree-hash_1.in
@@ -1,5 +1,10 @@
-4 
-4 0 1 1 2 
-4 2 0 2 3 
-4 0 1 1 1 
-4 0 1 2 3 
+10
+1 2
+1 3
+2 4
+2 5
+3 6
+3 7
+3 8
+8 9
+8 10
