+++ docs/graph/examples/tree-hash/tree-hash_3.ans
@@ -1,3 +1,4 @@
-2
-4
-8
+1 
+1 
+3 
+1
