+++ docs/graph/examples/tree-hash/tree-hash_3.in
@@ -1,15 +1,5 @@
-3
-4
-1 3
-2 3
-4 3
-5
-1 2
-2 3
-3 4
-4 5
-5
-1 2
-2 3
-3 4
-3 5
+4 
+4 0 1 1 2 
+4 2 0 2 3 
+4 0 1 1 1 
+4 0 1 2 3
