+++ docs/graph/examples/tree-hash/tree-hash_4.ans
@@ -1,1 +1,3 @@
+2
 4
+8
