+++ docs/graph/examples/tree-hash/tree-hash_4.in
@@ -1,10 +1,15 @@
-10
-1 2
+3
+4
 1 3
-2 4
-2 5
-3 6
-3 7
-3 8
-8 9
-8 10
+2 3
+4 3
+5
+1 2
+2 3
+3 4
+4 5
+5
+1 2
+2 3
+3 4
+3 5
